auth-service/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
sqlalchemy==2.0.35
psycopg[binary]==3.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
requests==2.32.3
authlib==1.3.2
python-dotenv==1.0.1
email-validator==2.2.0
httpx==0.27.2
itsdangerous==2.2.0